Updated packages configuration.
This commit is updating the package references that is needed by the
project. This commit also comments out 4 test cases that was already
ignored due to the syntax for ignoring in nunit 3 has been changed.
Can't seem to find documentation for Nunit 3 yet, so for now the test
cases are commented out using //